trace-cmd: Fix build dependencies
After having trace-cmd break because not all proper files are being compiled
when needed, I looked into why the dependencies were not working. I found
that the .foo.d files that were being created required having the full path
of the object file in them. Without that, the dependencies were being
ignored.

By adding "-MT $(bdir)/$*.o", to the dependency creating rules, the full
path would be added to targets for the object files, and this makes the
depencies work again.

Link: https://lore.kernel.org/linux-trace-devel/20200602181618.737914a8@oasis.local.home

Fixes: 60f0a40bb ("trace-cmd: Make libtraceevent buildable out-of-tree")
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>