tracefs events: Do not process str_read_file() if size is zero
If a file has no size (nothing is read), then str_read_file() frees the
buffer and returns zero. The problem is that all callers of str_read_file()
uses the buffer supplied if the value returned is not a negative. This
causes the freed buffer being used by the callers if the file read existed
but had no content.

This is apparent when using a copy of the tracefs directory, where some file
exist, but have no content, then loading the events would cause a segfault.

Change the callers to check the return value of str_read_file() for zero or
negative, and do not go further if it is.

Link: https://lore.kernel.org/linux-trace-devel/20201123154607.5e43e1ff@gandalf.local.home

Acked-by: "Tzvetomir Stoyanov (VMware)" <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>