libtracefs: Add $(LIBS) to building of shared object library
libtracefs.so relies on libtraceevent.so, but it an application were to
include libtracefs.so, it should not be required to include libtraceevent.so
if that application is not using it. Add the libtraceevent library to the
build of libtracefs.so so that what libtracefs.so uses is included in that
code.

Link: https://lore.kernel.org/linux-trace-devel/20201203190741.174872816@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>