libtracefs: Load ftrace events in fill_local_events_systems()
While writing some code that used tracefs to access the tracefs directory,
and to load the tep events, I found that it was missing the ftrace events. I
enabled the userstacktrace option, but found that the tep handler returned
by tracefs_local_events_system() did not contain any ftrace event, and thus
I could not parse the user_stack events.

This was due to the fill_local_events_systems() checking against the list of
system names returned by tracefs_event_systems(), which returns the names of
all the system directories that have an "enable" file in them. As ftrace
events are special, and the directory is only for parsing purposes (not to
enable events), it was skipped due to the lack of an "enable" file.

Instead, after processing all the systems returned by tracefs_event_systems(),
check if the sys_names is NULL or matches "ftrace", and if it does, add the
ftrace events as well.

Link: https://lore.kernel.org/linux-trace-devel/20201204170119.220692d8@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>