libtracefs: Fix soft links to have a proper order
Currently the libraries created have both the .so and the .so.X soft links
pointing to the full verision .so.X.Y.Z. It's better to have them point
serially instead of directly, as it is more common and also makes it easier
to change the .so to point to a different .so.X if needed.

Now the links will be:

  .so -> .so.X
  .so.X -> .so.X.Y.Z

Also clean up the dependencies a little in the Makefile.

Link: https://lore.kernel.org/linux-trace-devel/20201211103018.5f49ae1c@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>