libtracefs: Add trace.conf for ld.so if needed
It cannot be guaranteed that the system the library is being installed on
has ld.so set to search the location of where libtracefs.so is installed.

During the install process, check if the path is found in ld.so.conf.d/ and
if not, test if an executable can be built and find the library. If it can
not, then add the path to ld.so.conf.d/trace.conf, and run ldconfig again.

Link: https://lore.kernel.org/linux-trace-devel/20201215112807.5b99788a@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>