libtracefs: Remove tracefs_find_tracing_dir() API
There are two APIs to get the tracefs mount point:
tracefs_get_tracing_dir() and tracefs_find_tracing_dir().
The only difference between them is that the first returns
a static cached string and the second a newly allocated string.
In order not to confused the users with too many APIs that do
the same, the tracefs_find_tracing_dir() is removed as an API,
renamed to trace_find_tracing_dir() and is used only as a library
internal function.

Link: https://lore.kernel.org/linux-trace-devel/20201221072921.210230-2-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>