libtracefs: Fix notations of tracefs_function_filter() and module parameter
Highlight the function name and encapsulate each '*' with a "pass:[*]" to
not have the creation of the man pages interpret them as bold annotations.

Also remove the tracefs_instance_free(), as tracefs_instance_destroy() is
called and that frees the instance.

The description in the man page for tracefs_function_filter() left out
that NULL is a suitable value for the modules parameter, meaning that all
functions will be examined for a match of one of the filters.

Link: https://lore.kernel.org/linux-trace-devel/20210330005247.685321870@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>