libtracefs: Add checking of available_filter_functions to tracefs_function_filter()
Instead of modifying the set_ftrace_filter file with a filter that
possibly does not match anything, check the available_filter_functions
file first to make sure that each filter will match something. Then if
"reset" is enabled, it wont clear the filter file and not set any
filtering, which could cause issues if the application is not expecting
all functions to be enabled.

Link: https://lore.kernel.org/linux-trace-devel/20210330005248.103947204@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>