libtracefs: Allow for setting filters with regex expressions
All for full "regex(3)" processing of setting functions in the
set_ftrace_filter file. Check if the filter passed in is just a glob
expression that the kernel can process, or if it is a regex that should look
at the available_filter_functions list instead.

If it is a regex, it will read the available_filter_functions and write in
each function as it finds it.

Link: https://lore.kernel.org/linux-trace-devel/20210330005248.391852830@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>