libtracefs: Add indexing to set functions in tracefs_function_filter()
Starting in Linux kernel 5.1, it is possible to write the index of the
available_filter_functions (where the first function is 1), which is a
much faster way to set functions than with the glob expressions. By
finding the matches of the available_filter_functions and using the index
of the function, this can make the function call of
tracefs_function_filter() return much faster.

Link: https://lore.kernel.org/linux-trace-devel/20210330005248.543285809@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>