libtracefs: Just past one filter in for tracefs_function_filter()
With the new CONTINUE flag, it is inconsistent to have an array of filters
but only one module. Change the API to pass in a single filter, and remove
the errs parameter. Also change the return value to be 0 on success, 1 if
it failed before starting to write to the filter file, and -1 if it failed
after starting to write to the filter file.

Also, set errno appropriately, where if the filter fails EINVAL is set.

Link: https://lore.kernel.org/linux-trace-devel/20210329210302.0f7910e3@oasis.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>