libtracefs: Only allow RESET flag if file is not already opened
Only allow the flags to have TRACEFS_FL_RESET set in
tracefs_function_filter() if the previous call with the same instance did
not have the TRACEFS_FL_CONTINUE set.

The RESET flag means to reset the filter, and since it can only do that
when the file is opened, it does not make sense to have it set. It also is
dangerous to simply close (and commit previous changes) if it is set and
the file is already opened. The only rational solution is to fail the
call in this case.

Link: https://lore.kernel.org/linux-trace-devel/20210330105250.0a7105a1@gandalf.local.home/
Link: https://lore.kernel.org/linux-trace-devel/20210330183545.906539717@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>