libtracefs: Add TRACEFS_FL_FUTURE flag for future module filtering
Starting in Linux v4.13, it is possible to load filters of a function for
a module before it is loaded. That is:

 # echo '*:mod:rfkill' > set_ftrace_filter
 # echo function > current_tracer
 # modprobe rfkill

And the above will enable all functions for the rfkill module to be traced
before it starts to run.

Add a TRACEFS_FL_FUTURE flag to be able to accomplish the same thing.

Link: https://lore.kernel.org/linux-trace-devel/20210330183546.351496716@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>