libtracefs: Fix make install updating of ldconfig
The ldconfig file /etc/ld.conf.d/trace.conf was being updated when it did
not need to be, and should not be updated if ldconfig is not in the patch
when installing.

Add a check to see if ldconfig can successfully run before testing if conf
file should be updated. Also fix the grep of the conf file that looks for
the path such that the output is not displayed to the console.

Link: https://lore.kernel.org/linux-trace-devel/20210408211736.0f9cb218@oasis.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>