libtracefs: Have the option mask function return a const pointer
As the pointer returned by the tracefs_options_get_*() point to the
instance field, it must not be freed or modified by the caller. Have the
return value of the mask be a constant, and update the documentation to
reflect this.

Also fixed the documentation about supported options. If an option is
supported in the top level instance, it may *not* be supported in a
created instance. Currently all options supported in the created instance
are also supported at the top level, but there's no guarantee that will be
the case in the future. All created instances (mkdir in instances
directory) have the same supported options.

Link: https://lore.kernel.org/linux-trace-devel/CAPpZLN7G4AttMVFhwtf_YVouRtXhtNyHBK7QqEZWzAEeGhU_9Q@mail.gmail.com/
Link: https://lore.kernel.org/linux-trace-devel/20210409184930.2241303-2-rostedt@goodmis.org

Suggested-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>