libtracefs: Add loading of mappings in fill_local_events_system()
When writing some code for parsing, I found that tracefs_local_events()
did not load the necessary information to have the printing of events work
just right. It left out reading kallsyms, saved_cmdlines, and
printk_formats, which are necessary for several events to print properly.

Have fill_local_events_system() read and load those files such that the
events can be printed fully. If there's a problem with reading and loading
any of these, it does not make the function called failed. This is a best
effort approach.

Link: https://lore.kernel.org/linux-trace-devel/20210410003727.2288762-4-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>