libtracefs: Free the allocated kbuffer in tracefs_interate_raw_events()
When reading the ring buffers, a kbuffer is allocated and assigned to be
able to parse out the raw events from the kernel buffer. But it is never
freed and causes leaked memory.

Consolidate page_to_kbuf() into read_next_page(), and instead of
allocating the kbuffer descriptor each time, simply reload it with the new
page.

Link: https://lore.kernel.org/linux-trace-devel/20210412162838.544b3265@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>