libtracefs utest: Make a null string fatal if it is passed to strstr
strstr will segv for a null string. If a string is passed to strstr and
is null there is no point segving, it is better to fail the earlier test
with a fatal and continue the rest of the suite. This allows the suite
to complete with an error rather than die.

Link: https://lore.kernel.org/linux-trace-devel/20210423035438.2221760-3-irogers@google.com

Acked-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>
Signed-off-by: Ian Rogers <irogers@google.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>