libtracefs: Implement tracefs_filter_functions()
Add a new function that allows the user to get a list of functions that are
available to be filtered. This is useful to see if a function is available
before trying to add a kprobe to it. It can also be used to see what
architecture the code is being run on.

Link: https://lore.kernel.org/linux-trace-devel/20210615141541.08685845@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>