libtracefs: Check tracer parameter first in tracefs_tracer_set()
There's no reason to open the current_tracer file, nor even find its path
before checking if the tracer parameter is valid or not. Instead of
checking that later in the process and then having to "clean up" what has
already been done, check it first, and simply return with error if it is
not valid.

Link: https://lore.kernel.org/linux-trace-devel/20210625221739.1215557-4-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>