libtracefs: Implement tracefs_get_kprobes()
Add the function tracefs_get_kprobes() that returns a list of kprobes that
are registered. The list contains strings that are of the format
"group/event" (i.e. "kprobes/open"). The last element in the list is a
NULL pointer. In the case that there are no kprobes, the list will return
a single entry of a NULL pointer.

tracefs_get_kprobes() takes a type parameter to determine which kprobes to
return:

  TRACEFS_ALL_KPROBES	- return all that are found
  TRACEFS_KPROBE	- return only normal kprobes ("p:")
  TRACEFS_KRETPROBE	- return only kretprobes ("r:")

NULL is returned in the case of error (mainly a memory issue, or bad
parsing).

Link: https://lore.kernel.org/linux-trace-devel/20210702201727.169080-5-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>