libtracefs: Implement tracefs_kprobe_info()
Add tracefs_kprobe_info() that returns the type of a given probe, and may
also return the content of the parsed kprobe_events file.

If @type is non NULL, it will hold the content of the kprobe before the
":'.

If @addr is non NULL, it will hold the content of the kprobe's address or
function that it is attached to.

If @format is non NULL, it will hold the format string of the kprobe.

It returns:

    TRACEFS_ALL_KPROBES if an error occurs or the kprobe is not found,
             or the probe is of an unknown type.
             Having @type set, can help debug an unknown type.
    TRACEFS_KPROBE if the type of kprobe found is a normal kprobe.
    TRACEFS_KRETPROBE if the type of kprobe found is a kretprobe.

Link: https://lore.kernel.org/linux-trace-devel/20210702201727.169080-7-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>