libtracefs: Implement tracefs_kprobe_clear_all() to remove all kprobes
A call to tracefs_kprobe_clear_all() will attempt to disable all kprobes. If
any kprobe is set, and the @force parameter is set, it will fail with
errno set to EBUSY. If @force is set, then it will attempt to disable all
the defined kprobe events and then clear it.

Link: https://lore.kernel.org/linux-trace-devel/20210702201727.169080-8-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>