libtracefs: Add tracefs_iterate_stop()
In the API for data streaming we implemented a special
function (tracefs_trace_pipe_stop()) that can be called from
the user in order to terminate the continuous streaming of the
tracing data. Here we add similar functionality that can be
used to terminate the continuous iteration over the raw events.

Link: https://lore.kernel.org/linux-trace-devel/20210705122706.146428-1-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>