libtracefs: Remove "traceevent/" from referencing libtraceevent headers
"pkg-config --cflags --libtraceevent" produces the full path to the
libtraceevent headers (including the "traceevent/" portion). There is no
need to add that part in the headers.

Particularly the one in the main tracefs.h that trace-cmd uses. The
version 2.9.4 causes the header to include the libtraceevent/event-parse.h
that is local, when it should be using the system headers, and causes it
to fail the build.

Link: https://lore.kernel.org/linux-trace-devel/20210716220649.46aab3f7@rorschach.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>