libtracefs: Add tracefs_list_pop() to remove the last item
Add a way to remove the last item of a string list. This will be needed to
revert changes made to the new tracefs_synth structure that will have more
than one tracefs_list char ** that must be the same size. If one is
successfully updated, but the next one fails, the first one needs to be
reverted.

Link: https://lore.kernel.org/linux-trace-devel/20210722133610.379933-2-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>