libtracefs: Add Requires libtraceevent to pkg-config file
If a developer uses libtracefs only, they should not need to know about
libtraceevent just because libtracefs requires it.

The pkg-config pc file has an option to handle such cases, called
"Requires:". Utilize this feature so that users using libtracefs only need
to do:

  CFLAGS = `pkg-config --cflags --libs libtracefs`

And not:

  CFLAGS = `pkg-config --cflags --libs libtracefs libtraceevent`

As with the Requires option, pkg-config now produces:

 $ pkg-config --cflags --libs libtracefs
 -I/usr/local/include/tracefs -I/usr/local/include/traceevent -ltracefs -ltraceevent

Link: https://lore.kernel.org/linux-trace-devel/20210722155435.773e91ec@oasis.local.home

Reported-by: Daniel Bristot de Oliveira <bristot@redhat.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>