libtracefs: Update to version 1.3-dev
Since the development of 1.3 source has diverged from the stable 1.2 path,
it still has the old 1.2.1 as the version number. If this is added to a
system with 1.2.2 or above, it wont be used, as the symlinks point to the
highest version (1.2.2 or above).

Set this to 1.3.dev to make it the latest version to be used.

Link: https://lore.kernel.org/linux-trace-devel/20210730030307.574270-2-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>