libtracefs: Add files created by bison and flex
The files sqlhist.tab.c, sqlhist.tab.h and sqlhist-lex.c are dynamically
created by flex and bison when their source files change (sqlhist.y and
sqlhist.l). But these seldom do change, and there's no reason to have
the build require bison and flex if people are only trying to build the
code and not modify it. Add the created files into the repo so that flex
and bison are not required to do the build.

Link: https://lore.kernel.org/linux-trace-devel/20210825195849.7f2d35d8@rorschach.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>