libtracefs: Transform tracefs_hist_add_sort_key()
The current version of the API makes it hard to add multiple sort keys
to a histogram. The only way to do this is to use the variadic arguments,
however in order to do this the caller has to know the number of sort
keys at compile time, because the method overwrite all previously added
keys. The problem is addressed by creating a tracefs_hist_add_sort_key()
into two methods - one that overwrite and one that does not.
The current version of 'tracefs_hist_add_sort_key()' gets renamed to
'tracefs_hist_set_sort_key()' without introducing any functional changes.
In the same time a new 'tracefs_hist_add_sort_key()' function is
defined. The new function can add one new sort key to the list of
previously added sort keys.

Link: https://lore.kernel.org/linux-trace-devel/20210924095702.151826-3-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>